# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chinook API",
        "contact": {
            "name": "Testsmith",
            "email": "info@testsmith.io"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local environment"
        }
    ],
    "paths": {
...
        "/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Retrieve all customers",
                "operationId": "get-all-customers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page, default=1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "New customer",
                "operationId": "store-customer",
                "requestBody": {
                    "description": "Customer object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/login": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Login customer",
                "operationId": "login-customer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "",
                                            "type": "string",
                                            "example": ""
                                        },
                                        "token_type": {
                                            "description": "",
                                            "type": "string",
                                            "example": ""
                                        },
                                        "expires_in": {
                                            "description": "",
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/me": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Retrieve current customer info",
                "operationId": "get-current-customer-info",
                "responses": {
                    "200": {
                        "description": "A customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/customers/logout": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Destroy token",
                "operationId": "destroy token",
                "responses": {
                    "200": {
                        "description": "A token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "",
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/customers/{customerId}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Retrieve a specific customer",
                "operationId": "get-customer",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "The customerId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Update customer",
                "operationId": "update-customer",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "The customerId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Customer object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Remove customer",
                "operationId": "destroy-customer",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "The customerId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/customers/search": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Retrieve specific customers matching the search query",
                "operationId": "search-customer",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The query parameter in path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All matching customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerRequest": {
                "title": "CustomerRequest",
                "required": [
                    "firstname",
                    "lastname",
                    "address",
                    "city",
                    "country",
                    "email"
                ],
                "properties": {
                    "firstname": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "company": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "address": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "city": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "state": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "country": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "postalcode": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "fax": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "email": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "password": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "support_rep_id": {
                        "description": "Class Customer",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CustomerResponse": {
                "title": "CustomerResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "postalcode": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "fax": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "support_rep_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "support_rep": {
                        "$ref": "#/components/schemas/EmployeeResponse"
                    }
                },
                "type": "object"
            },
...
            "EmployeeResponse": {
                "title": "EmployeeResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "reports_to": {
                        "$ref": "#/components/schemas/EmployeeResponse"
                    },
                    "birthdate": {
                        "type": "string"
                    },
                    "hiredate": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "postalcode": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "fax": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "customers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CustomerResponse"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}